requirements.txt
Django==6.0.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
django-allauth==65.14.0
django-crispy-forms==2.5
crispy-bootstrap5==2025.6
django-widget-tweaks==1.5.1
requests==2.32.5
PyJWT==2.10.1
cryptography==46.0.4
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
idna==3.11
pycparser==3.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
